backend/requirements.txt
Flask
python-dotenv        
openai             
flask-sqlalchemy   
psycopg2-binary    
flask-migrate       
Flask-WTF           
WTForms             
Flask-CORS
pytest
pytest-flask
flask-bcrypt
boto3
botocore
mimetype